+++ bindings/python/TSID/include/BipedalLocomotion/bindings/TSID/VariableFeasibleRegionTask.h
@@ -1,5 +1,5 @@
 /**
- * @file VariableFeaibleRegionTask.h
+ * @file VariableFeasibleRegionTask.h
  * @authors Roberto Mauceri
  * @copyright 2025 Istituto Italiano di Tecnologia (IIT). This software may be modified and
  * distributed under the terms of the BSD-3-Clause license.