+++ src/TSID/src/VariableFeasibleRegionTask.cpp
@@ -5,8 +5,8 @@
  * distributed under the terms of the BSD-3-Clause license.
  */
 
-#include <BipedalLocomotion/TSID/VariableFeasibleRegionTask.h>
 #include <BipedalLocomotion/System/VariablesHandler.h>
+#include <BipedalLocomotion/TSID/VariableFeasibleRegionTask.h>
 #include <BipedalLocomotion/TextLogging/Logger.h>
 
 #include <iDynTree/EigenHelpers.h>
@@ -24,7 +24,8 @@ bool VariableFeasibleRegionTask::setVariablesHandler(const VariablesHandler& var
 
     if (!m_isInitialized)
     {
-        log()->error("{} The task is not initialized. Please call initialize() method.", errorPrefix);
+        log()->error("{} The task is not initialized. Please call initialize() method.",
+                     errorPrefix);
         return false;
     }
 
@@ -38,7 +39,8 @@ bool VariableFeasibleRegionTask::setVariablesHandler(const VariablesHandler& var
 
     if (m_variableSize > variable.size)
     {
-        log()->error("{} The expected size is greater than the one stored in the variable named {}.",
+        log()->error("{} The expected size is greater than the one stored in the variable named "
+                     "{}.",
                      errorPrefix,
                      m_variableName);
         return false;
@@ -65,8 +67,7 @@ bool VariableFeasibleRegionTask::setVariablesHandler(const VariablesHandler& var
             }
             m_S(i, index) = 1;
         }
-    } 
-    else
+    } else
     {
         if (m_variableSize != variable.size)
         {
@@ -80,7 +81,8 @@ bool VariableFeasibleRegionTask::setVariablesHandler(const VariablesHandler& var
         }
         // if the size of the m_controlledElements vector is zero, this means that the entire
         // variable is regularized
-        // iDynTree::toEigen(this->subA(variable)).setIdentity(); // devo sostituirlo col comando sotto?
+        // iDynTree::toEigen(this->subA(variable)).setIdentity(); // devo sostituirlo col comando
+        // sotto?
         m_S = Eigen::MatrixXd::Identity(m_variableSize, m_variableSize);
     }
     return true;
@@ -88,7 +90,7 @@ bool VariableFeasibleRegionTask::setVariablesHandler(const VariablesHandler& var
 
 bool VariableFeasibleRegionTask::initialize(std::weak_ptr<const IParametersHandler> paramHandler)
 {
-   constexpr auto errorPrefix = "[VariableFeasibleRegionTask::initialize]";
+    constexpr auto errorPrefix = "[VariableFeasibleRegionTask::initialize]";
 
     auto ptr = paramHandler.lock();
     if (ptr == nullptr)
@@ -143,17 +145,17 @@ bool VariableFeasibleRegionTask::initialize(std::weak_ptr<const IParametersHandl
     return true;
 }
 
-bool VariableFeasibleRegionTask::setFeasibleRegion(
-    Eigen::Ref<const Eigen::MatrixXd> C, 
-    Eigen::Ref<const Eigen::VectorXd> l, 
-    Eigen::Ref<const Eigen::VectorXd> u)
+bool VariableFeasibleRegionTask::setFeasibleRegion(Eigen::Ref<const Eigen::MatrixXd> C,
+                                                   Eigen::Ref<const Eigen::VectorXd> l,
+                                                   Eigen::Ref<const Eigen::VectorXd> u)
 {
     m_isValid = false;
     constexpr auto errorPrefix = "[VariableFeasibleRegionTask::setFeasibleRegion]";
     // check if the size of the matrices is correct
     if (C.cols() != m_variableSize)
     {
-        log()->error("{} The number of columns of the matrix C is not correct. Expected: {}, Passed: {}.",
+        log()->error("{} The number of columns of the matrix C is not correct. Expected: {}, "
+                     "Passed: {}.",
                      errorPrefix,
                      m_variableSize,
                      C.cols());
@@ -178,7 +180,7 @@ bool VariableFeasibleRegionTask::setFeasibleRegion(
     if (!(u.array() >= l.array()).all())
     {
         log()->error("{} The elements of the vector u must be greater than or equal to"
-        " the elements of the vector l.",
+                     " the elements of the vector l.",
                      errorPrefix);
         return false;
     }